Move process code to separate package
